+++ app/src/main/java/com/knowlily/browser/model/BookmarkItem.kt
@@ -1,3 +1,12 @@
 package com.knowlily.browser.model
 
-data class BookmarkItem(val url: String, val timestamp: Long)
+import androidx.room.ColumnInfo
+import androidx.room.Entity
+import androidx.room.PrimaryKey
+
+@Entity(tableName = "bookmarks")
+data class BookmarkItem(
+    @PrimaryKey(autoGenerate = true) val id: Long = 0,
+    @ColumnInfo(name = "url") val url: String,
+    @ColumnInfo(name = "timestamp") val timestamp: Long
+)